--- app/src/main/kotlin/de/timklge/karooheadwind/Throttle.kt
@@ -1,16 +1,0 @@
-package de.timklge.karooheadwind
-
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.flow
-
-fun<T> Flow<T>.throttle(timeout: Long): Flow<T> = flow {
-    var lastEmissionTime = 0L
-
-    collect { value ->
-        val currentTime = System.currentTimeMillis()
-        if (currentTime - lastEmissionTime >= timeout) {
-            emit(value)
-            lastEmissionTime = currentTime
-        }
-    }
-}
