+++ app/src/main/kotlin/de/timklge/karooheadwind/Utils.kt
@@ -1,0 +1,21 @@
+package de.timklge.karooheadwind
+
+import kotlin.math.abs
+
+fun signedAngleDifference(angle1: Double, angle2: Double): Double {
+    val a1 = angle1 % 360
+    val a2 = angle2 % 360
+    var diff = abs(a1 - a2)
+
+    val sign = if (a1 < a2) {
+        if (diff > 180.0) -1 else 1
+    } else {
+        if (diff > 180.0) 1 else -1
+    }
+
+    if (diff > 180.0) {
+        diff = 360.0 - diff
+    }
+
+    return sign * diff
+}
